+++ crates/miden-testing/src/kernel_tests/tx/test_account_delta.rs
@@ -789,6 +789,27 @@ fn asset_and_storage_delta() -> anyhow::Result<()> {
     Ok(())
 }
 
+/// Tests that adding a fungible asset with amount zero to the account vault works and does not
+/// result in an account delta entry.
+#[test]
+fn adding_amount_zero_fungible_asset_to_account_vault_works() -> anyhow::Result<()> {
+    let mut builder = MockChain::builder();
+    let account = builder.add_existing_mock_account(Auth::IncrNonce)?;
+    let input_note = builder.add_p2id_note(
+        account.id(),
+        account.id(),
+        &[FungibleAsset::mock(0)],
+        NoteType::Private,
+    )?;
+    let chain = builder.build()?;
+
+    let tx = chain.build_tx_context(account, &[input_note.id()], &[])?.build()?.execute()?;
+
+    assert!(tx.account_delta().vault().is_empty());
+
+    Ok(())
+}
+
 // TEST HELPERS
 // ================================================================================================